# [SPARK-36827][CORE] Improve the perf and memory usage of cleaning up stage UI data