@gengliangwang 2021-09-24
@mridulm there is something wrong with the benchmark. I rerun the updated benchmark and find that after changes the LevelDB is slower. 
I am considering making the code changes for InMemoryStore only. So I mark it as WIP.