@gengliangwang 2021-09-24
@mridulm Thanks for the review. This one is ready now.
As I observe regression after applying the changes when using LevelDB, this PR is to optimize InMemoryStore only.
Now the PR is safe enough to go into Spark 3.2 as well