@mridulm 2021-09-24
I am guessing the minor perf difference for level db would be the sort being done after pulling all data vs pushing the sort to the store.
One option is to batch all the deletes together, instead of doing it one by one - which should make level db updates also faster ?