@gengliangwang 2021-09-24
> One option is to batch all the deletes together, instead of doing it one by one - which should make level db updates also faster ?

Then the same optimization can be done in the `cleanupStagesInKVStore` as well and the minor perf difference is still there. We can try that in another PR.