@mridulm 2021-09-24
Btw, can you share the details of the regression ?
What I had seen earlier was:
```
Before changes:
InMemoryStore: 2.8s
LevelDB: 68.9s

After changes:
InMemoryStore: 0.95s
LevelDB: 60.3s
```

That looked fine to me - did we observe a degradation in timing results after ?