@gengliangwang 2021-09-24
@mridulm Yes. 
At first the benchmark of job start:
```
    listener.onJobStart(SparkListenerJobStart(4, time, stages, null))
```

Later on, I change the stage info inside to `Nil`
```
    listener.onJobStart(SparkListenerJobStart(4, time, Nil, null))
```
and the difference for LevelDB is
before changes: 2.5s
after changes: 3s