@zhouyejoe 2021-09-25
@gengliangwang 
QQ regarding the regression test:

> the difference for LevelDB is
> before changes: 2.5s
> after changes: 3s

How many stages within your test log file for SHS? Or will this performance regression get amplified if there are more stages within the application?