@gengliangwang 2021-09-25
@zhouyejoe I use the benchmark code in the PR description, which is runnable under `AppStatusListenerSuite`.
I tried 3000 and 5000 stages the result "after changes" is always slightly slow. 
When the number of stages reaches 1001, Spark will delete the oldest 100 stages UI data. I guess after changes reading 1001 stages from LevelDB is slow. Before changes Spark only reads 100 stages from LevelDB.